Move fields test into fields package
